+++ apps/website/src/app/services/old/chromaEffect/state/multi-dynamic-state/multi-dynamic-state.ts
@@ -1,6 +1,6 @@
 import { State } from '../abstract/state';
-import { color_wheel, randomInteger } from '../../../../shared/functions';
-import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../old/chromaSDK/old-chroma-s-d-k.service';
+import { color_wheel, randomInteger } from '../../../../../shared/functions';
+import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../chromaSDK/old-chroma-s-d-k.service';
 
 export class MultiDynamicState extends State {
   private headsetColors = [0, 0, 0, 0, 0];
