+++ apps/website/src/app/services/old/chromaEffect/state/waterfall-state/waterfall-state.ts
@@ -1,12 +1,12 @@
 import { State } from '../abstract/state';
-import { calculateBGRInteger, mapNumber, qadd8, qsub8, randomInteger, WarmColor } from '../../../../shared/functions';
+import { calculateBGRInteger, ColdColor, mapNumber, qadd8, qsub8, randomInteger } from '../../../../../shared/functions';
 import iro from '@jaames/iro';
-import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../old/chromaSDK/old-chroma-s-d-k.service';
+import { HeadsetEffect, KeyboardEffect, MouseEffect } from '../../../chromaSDK/old-chroma-s-d-k.service';
 
-export class Fire2012State extends State {
+export class WaterfallState extends State {
   readonly COOLING = 55;
   readonly SPARKING = 120;
-  private interval: string | number | NodeJS.Timer | undefined;
+  private interval!: NodeJS.Timeout;
   private G_REVERSE_DIRECTION = false;
   private heatKeyboard: number[] = [];
   private heatMouse: number[] = [];
@@ -18,8 +18,8 @@ export class Fire2012State extends State {
   handle(colors: iro.Color[]): void {
     if (this.interval === undefined) {
       this.interval = setInterval(() => {
-        this.fire2012Keyboard();
-        this.fire2012Mouse();
+        this.waterfallKeyboard();
+        this.waterfallMouse();
 
         const headsetColors = this._context.keyboardColors[0].splice(0, this._context.headset.amount);
         this._context.createHeadsetEffect(HeadsetEffect.CHROMA_CUSTOM, headsetColors);
@@ -44,41 +44,7 @@ export class Fire2012State extends State {
     clearInterval(this.interval);
   }
 
-  // Fire2012 by Mark Kriegsman, July 2012
-  // as part of "Five Elements" shown here: http://youtu.be/knWiGsmgycY
-  // This basic one-dimensional 'fire' simulation works roughly as follows:
-  // There's a underlying array of 'heatKeyboard' cells, that model the temperature
-  // at each point along the line.  Every cycle through the simulation,
-  // four steps are performed:
-  //  1) All cells cool down a little bit, losing heatKeyboard to the air
-  //  2) The heatKeyboard from each cell drifts 'up' and diffuses a little
-  //  3) Sometimes randomly new 'sparks' of heatKeyboard are added at the bottom
-  //  4) The heatKeyboard from each cell is rendered as a color into the leds array
-  //     The heatKeyboard-to-color mapping uses a black-body radiation approximation.
-  //
-  // Temperature is in arbitrary units from 0 (cold black) to 255 (white hot).
-  //
-  // This simulation scales it self a bit depending on LED_COUNT; it should look
-  // "OK" on anywhere from 20 to 100 LEDs without too much tweaking.
-  //
-  // I recommend running this simulation at anywhere from 30-100 frames per second,
-  // meaning an interframe delay of about 10-35 milliseconds.
-  //
-  // Looks best on a high-density LED setup (60+ pixels/meter).
-  //
-  //
-  // There are two main parameters you can play with to control the look and
-  // feel of your fire: COOLING (used in step 1 above), and SPARKING (used
-  // in step 3 above).
-  //
-  // COOLING: How much does the air cool as it rises?
-  // Less cooling = taller flames.  More cooling = shorter flames.
-  // Default 50, suggested range 20-100
-
-  // SPARKING: What chance (out of 255) is there that a new spark will be lit?
-  // Higher chance = more roaring fire.  Lower chance = more flickery fire.
-  // Default 120, suggested range 50-200.
-  private fire2012Keyboard(): void {
+  private waterfallKeyboard(): void {
     this.coolDown(this.heatKeyboard, this._context.keyboard.columns);
     this.driftUp(this.heatKeyboard, this._context.keyboard.columns);
 
@@ -87,7 +53,7 @@ export class Fire2012State extends State {
     // Step 4.  Map from heatKeyboard cells to LED colors on keyboard
     for (let i = 0; i < this._context.keyboard.rows; i++) {
       for (let j = 0; j < this._context.keyboard.columns; j++) {
-        const color = WarmColor(mapNumber(this.heatKeyboard[j], 0, 255, 0, 240));
+        const color = ColdColor(mapNumber(this.heatKeyboard[j], 0, 255, 0, 240));
         let pixelnumber;
         if (this.G_REVERSE_DIRECTION) {
           pixelnumber = (this._context.keyboard.columns - 1) - j;
@@ -107,7 +73,7 @@ export class Fire2012State extends State {
     this._context.createKeyboardEffect(KeyboardEffect.CHROMA_CUSTOM, this._context.keyboardColors).then();
   }
 
-  private fire2012Mouse(): void {
+  private waterfallMouse(): void {
     this.coolDown(this.heatMouse, this._context.mouse.rows);
     this.driftUp(this.heatMouse, this._context.mouse.rows);
 
@@ -115,7 +81,7 @@ export class Fire2012State extends State {
     // Step 4.  Map from heatKeyboard cells to LED colors on keyboard
     for (let i = 0; i < this._context.mouse.rows; i++) {
       for (let j = 0; j < this._context.mouse.columns; j++) {
-        const color = WarmColor(mapNumber(this.heatMouse[i], 0, 255, 0, 240));
+        const color = ColdColor(mapNumber(this.heatMouse[i], 0, 255, 0, 240));
         let pixelnumber;
         if (this.G_REVERSE_DIRECTION) {
           pixelnumber = (this._context.mouse.rows - 1) - i;
